rename Dto to QueryResult in Meta BC
