+++ src/Core/Domain/Meta/Query/GetPagesForLayoutCustomization.php
@@ -27,7 +27,7 @@
 namespace PrestaShop\PrestaShop\Core\Domain\Meta\Query;
 
 /**
- * Class GetPagesForLayoutCustomization gets pages which layout can be customized.
+ * Gets pages for which theme's layout can be customized.
  */
 class GetPagesForLayoutCustomization
 {
