+++ src/PrestaShopBundle/Form/Admin/Improve/Design/Theme/PageLayoutCustomizationFormFactoryInterface.php
@@ -26,7 +26,7 @@
 
 namespace PrestaShopBundle\Form\Admin\Improve\Design\Theme;
 
-use PrestaShop\PrestaShop\Core\Domain\Meta\DataTransferObject\LayoutCustomizationPage;
+use PrestaShop\PrestaShop\Core\Domain\Meta\QueryResult\LayoutCustomizationPage;
 use Symfony\Component\Form\FormInterface;
 
 /**
